Day twelve.
Went to all the trouble to implement a (half-assed) Dijkstra (and some Hashable
boilerplate) only to discover it didn't improve over my naive algorithm,
presumably because the terrain is so weird.

Oh well. Who doesn't love to implement a binary search because the standard
library has no way to (efficiently) maintain a sorted list of numbers? Let
alone an efficient priority queue.