A bit faster and a lot more cumbersome.
I did get Instruments running and we were spending an awful lot of time in
ClosedRange.init. Is it not a struct? How complicated is it to initialize.
Anyhow, rolling our own range tracking gets us from ~13s to ~9s which is still
ridiculously slow, but now we're spending 65% of our time in
IndexingIterator.next. Dare I try to optimize that?