Manually iterating over sources array takes us from ~9s to ~4s.
Now, I'm a proponent of developer productivity over performance to be sure, but
when designing a new language, you gotta make sure that iterating over an array
is fast. Ain't nobody got time to be thinking about whether they're using the
"fast" iteration mechanism or the slow one.

Seeing this code makes me sad. I don't want to have to give up all of my toys
just to be able to do a few billion math operations in a civilized amount of
time. I'm not asking for cycle parity with optimized C code, just don't make me
throw an order of magnitude out the window just for using the "nice" dishware.

Anyhow, there is probably some clever way to math this problem and I should
just think about it for a while and shave off a few orders of magnitude in the
number of computations I need to do, but I still feel a bit let down by Swift.