Day seventeen.
Sloppy as hell, but I'll commit now and clean up later in the spirit of honesty
(with the exculpatory excuse that much of this was written while my children
were running around shouting at the top of their lungs -- not an environment
conducive to concentration).

At the very least I have to get rid of the embedded cycle position, which I
determined by inspecting the output while running on my personal input data (it
cycles at a different spot in the sample data). It's insufficiently general!