Day nineteen.
I managed a subpar solution for part one, but my search space pruning was not
up to the task for part two. Given that I'm on vacation and not torturing
myself, I decided to check out how others did it, and stole a couple of
sensible pruning tactics which got me over the hump.

The code is also more verbose than it likely needs to be, oh well. I was trying
to avoid array indirection and bounds checking costs, so I put everything into
structs with fixed fields.