String helper rethinks.
Swift Algorithms has a split() that works on strings, and which handles
omitting empty subsequences. So yay for that, no need for splitAndTrim.

Added an after(sep) method for skipping everything up to and including some
separator string.

And in order to work with String and Substring, we need to extend
StringProtocol and cope with some other rough edges around having Substring
instead of String. Meh.