Day 11.
Was baffled for a while why my answer to the second part was wrong before
realizing that in part 1 I was adding 1 to expand space 2x and so in part 2 I
should add 999999 to expand space 1000000x, _not_ add 1000000. Oops!

Also hat tip to Swift Algorithms for providing combinations(). Saved me an
annoyed nested reduce.