Day 12.
The initial solution was a bit hairy, but fine. Kooky tree search. I did not
intuit that memoization would be needed for the second part. Though the problem
was clearly recursive, it did not seem obvious that there would be a ton of
repeated spring & group suffixes. In retrospect, there were. I am dumb.

Again I am hindered by the fact that a tuple of Hashable parts is not Hashable
in Swift.