Day 13.
Not too bad, though I botched the "avoid repeating the original solution" check
and was just checking for a repeated result rather than passing in the
separation axis to be ignored. One has to pass in the axis to be ignored
because otherwise one might find that same axis again even though there's
another "later" axis of symmetry that's entirely hidden inside the "ignored
extra stuff" from the first axis.

Not going to code golf this one more because covid is back to sapping me of my
extra energy and enthusiasm. Yay.