Day 17.
I started with plain A* and tried various hacks to account for the max straight
path condition but was just not getting it. Then I tried Dikjstra but that
didn't work either. In the end, I had to look it up; I'm dumb.

Given the simple insight that we're searching a 4D space instead of a 2D space,
everything was much simpler and fell into place immediately with no funny
business.

I can try to blame it on the brutal sinus headache I have had the last three
days, but I probably would not have figured it out even had my brain been
working at normal capacity, so I won't flatter myself.