Day 20.
Part one was fun, make a little circuit emulator. Part two was dumb. Analyzing
the network algorithmically smacked of the halting problem, so that seemed out
of the question. I thought about maybe optimizing it or optimizing the
emulator, but those didn't bear fruit.

In the end, I had to look it up. The "answer" was "graph and visually inspect
the network, notice that there happens to be a four input Conjunction gate that
outputs to rx, run the network and observe that the inputs to that four input
Conjunction gate happen to cycle with some period, compute the LCM of those
periods." Unsatisfying.