Day 21.
Part one was straightforward. Part two was bananas.

There was _no way_ I was going to inspect the 131x131 ASCII grid that was my
input, happen to notice that it was constructed in just such a way that it
resulted in the step number growing as a simple quadratic (with x having the
special form 65+131k), and then realize that I could compute three points on
that quadratic and fit a curve to it. Riiiiight.

So yeah, I thought for a while about ways that one might solve the problem in
general, came up with nothing, looked up how other people did it, rolled my
eyes, and implemented the nice Lagrange interpolation polynomial that someone
much smarter than me posted about on Reddit.

At least the earlier days have mostly been fun. My expectation for the
remaining three days is further annoyance and frustration. Merry Christmas!