Day 23.
I did part one with an algorithm that operated on the raw grid, but that ran
pretty slowly, and when part two upped the ante by removing all the one-way
edges in the graph, I knew I needed something fancier.

After confirming that longest path is NP complete in general, I figured I had
better look at the data because this was probably not a case of me just not
knowing about the right algorithm.

As I suspected, the input is carefully constructed such that really we have a
simpler cyclic graph with weighted edges and about 35 nodes in it. That's much
more brute-forceable than a 141x141 graph of individual cells. So I wrote some
code to trace the ASCII grid into a weighted edge graph, and then did a depth
first traversal of that for the longest path.

Since I couldn't abide having two totally separate bits of code for part one
and two, I went back and made the graph extractor handle icy slopes by
eliminating the back edge in those cases.