Day 25.
At first I tried brute forcing it, which worked for the example input, but was
definitely not going to work for the full graph with 7000+ edges. Then for a
while I thought maybe this requires knowledge of some graph theory theorem (I
think this is some variant of the minimum cut problem), but I kept noodling on
it.

Eventually I came up with an idea for how to pare down the candidate edges for
cutting, but as I was testing that, I realized that I should be able to pare
them right down to the exact set of edges that needed to be cut.

The idea is pretty simple, for any given candidate edge (between A and B), take
all the neighbors of B, and see if they can reach any of the neighbors of A in
N hops, without ever passing across the A-B edge. It's fairly efficient to see
if a node is reachable from some other node in N hops when N is low (I maxed
out at 5), so it's pretty cheap to filter down the edges.

As it turned out, I couldn't quite filter to the exact set of edges to cut. I
pared it down to four edges instead of three, presumably because even though
you can't go back across one of the graph partitioning edges, you can go back
across one of the other two. But "brute force" checking all 3 size 3 subsets of
a size 4 set didn't break the bank. :)

Anyhow, I'm happy I kept thinking about it. Having been demoralized by a few of
the later days' part twos, I was tempted to throw in the towel and just see
what the kids were talking about on the Reddits, but I feel better for having
solved it entirely myself. Maybe I can hang on a few more years before they put
me out to pasture.