Day 11.
Actually had to think/do some math today.

The handling of the memo map is not pretty, but I can't use the (lovely) Entry
API because it needs to hold ownership of the memo map while we compute the new
value to be inserted. I need to mutate the crap out of the map while computing
that value, so I have to give up ownership (and rehash the key when I'm finally
ready to insert).