Day 12.
There's probably a more elegant way to compute the sides of a region, using
some geometry that I'm not conversant in. But after implementing my first idea,
which was to "trace out" the edge and count the sides as I went, and realizing
that oops, some regions enclose other regions and that technique will not
handle such regions, I decided to just go with this brute force method.

I enumerate all the fences on a given "line" and then count how many gaps there
are in between fences on that line to determine the number of sides on that
line (gaps plus one).

There may be a fancier way to enumerate the regions as well, or perhaps to
compute the fences as we're enumerating the regions and perhaps avoid some of
these intermediate sets, but eh, this works and is fairly readable.