Day 19.
I was not familiar with the "word break" problem/algorithm, so I beat my head
against this for a while. I finally realized that if a more complex towel could
be built from simpler towels, it could be ignored completely (for part one).
That, combined with an efficient trie-based pattern checker got the job done
for part one.

Then I saw that in part two, we actually want to count up the hundreds of
billions of ways that the patterns could be made. After beating my head against
that for a while, I realized that this must just be a well established
algorithm, so I asked ChatGPT what it was called and it told me.

I don't feel too bad about not rediscovering this dynamic programming algorithm
from scratch having no knowledge of it, but I am reminded that I should think
in a more dynamic programming-y way when doing AoC problems. Invariably they
will lean in that direction at least a couple of times in any given season.