Day 21.
Phew, that was fiddly. I managed to convince myself that subpaths could not
impact the length of the outer paths, but I was only thinking about that for
the recursive dpad to dpad case, not the npad to dpad case.

Once I finally realized that I needed to search all possible paths, the
refactoring needed to do that, plus the earlier thinking about the dpad to dpad
fixed-pointed-ness, made going from part 1 to part 2 pretty easy. I just had to
cache the lengths for all possible A to B pairs at each depth, which I knew was
safe to do.

I'm a couple of days behind now, due to fighting off a cold. I'm sure Santa
won't mind if I finish the last couple of days after Christmas.