Day 22.
Surprisingly uncomplicated. The bit twiddling was a nice little lark, and then
the search for the best price was straightforward.

It would have been even more straightforward had I realized in the beginning
that my cute idea of sorting the price sequences from highest to lowest was
invalid. A monkey will sell on the _first_ occurrence of a given sequence, so
we need them in their naturally occuring order.