Day 23.
My first solution to part two was not quite brute force, but also not very
sophisticated and took ten minutes to run. Assuming this was some well known
graph clique problem, I googled it, and indeed it's known to be NP complete,
and there are various standard algorithms.

However, I observed that if you consider all "triangles" coming from a given
computer (graphs of size three which are all connected, which we had to find
for the first part), the largest group had no "dangling triangles", which is to
say a group of three computers where only one or two were at the LAN party.

So I wrote a simple solution that just iterates over every computer, computes
all triangles in their connection set and then checks to see if that set is
fully connected. If so, it's a candidate max clique and then we just find the
largest of those.

Cheating? Maybe, but the example and my input data both had that structure.