Day 24.
Part two was not readily amenable to algorithmic solution. I did actually take
a crack and wrote a search algorithm which relied on the cascading structure of
a binary adder. It found the lowest bit that was incorrect, then looked at the
dependencies of that output bit and tried swapping them all. The expectation
was that something upstream of the incorrect bit would have to be out of place.

That didn't quite work. If the bit itself was out of place, it could have been
swapped with something that was not a dependency of that bit.

I wrote some code to generate a graph of the circuit and eyeballed the first
invalid bit and saw that it had indeed been swapped with something a bit
downstream of its own dependencies.

I didn't feel like bothering to write code to carve out the "neighborhood"
around an output bit to search for swap candidates. It was easy enough to grok
the general structure of the adder they were using and then visually scan the
graph to find the out of place nodes.

It's my party and I'll use my eyes if I want to.