AoC 2025 day 1.
Started doing it in Unison and ran into some inscrutable bullshit. Not in the
mood to sort that out at the moment, so I'm just going to do it in Scala and
then maybe I'll come back and redo everything in Unison later when my loins are
properly girded.

I'm already five days behind schedule, let's not start lining up yaks to be
shaved.