Day 10.
Well, that was an adventure. I managed to squeeze a solution out of my brain
that took 45 minutes to run and still had some bugs, and decided that I was
probably not likely to reach enlightenment on my own within the amount of time
I have reserved for AoC, so I went to the internets.

As I suspected, most people just used Z3, which is not my jam. The rest used
Gaussian elimination and some searching, which I should probably have tried
myself. I realized early on that the solution would basically be an
underspecified set of linear equations. But one guy found a clever and elegant
search strategy, the kind that warms my heart. I replaced my not very clever
search strategy with one based on his idea, and now I can move on in peace.